+++ decoders/jkbms_can/__init__.py
@@ -1,0 +1,5 @@
+'''
+JKBMS native CAN protocol decoder.
+'''
+
+from .pd import Decoder